+++ NEWS.md
@@ -1,3 +1,5 @@
+# tabula 3.1.0.9000
+
 # tabula 3.1.0
 ## New classes and methods
 * Add `she()` to plot SHE analysis.